app/requirements.txt
fastapi==0.31.0
uvicorn==0.8.4
python-multipart
passlib[bcrypt]
databases
pyjwt
pymongo